Modify test to reduce nested levels for tinygo
tinygo v0.33 fails with roughly 24+ levels of nested objects.

This commit modifies TestUnmarshalDeepNesting to reduce nested
levels to 24 when testing marshaling and unmarshaling deeply
nested object.
